+++ index.html
@@ -108,5 +108,6 @@
 <script src="https://www.gstatic.com/firebasejs/9.17.1/firebase-firestore.js" type="module"></script>
 <script src="./scripts/index.js" type="module"></script>
 <script src="./scripts/footer.js"></script>
+<script src="./scripts/cartWishlist.js" type="module"></script>
 
 </html>
