+++ scripts/index.js
@@ -2,7 +2,6 @@ import { db } from "./firestoreConfig.js";
 import { collection, getDocs, setDoc, doc } from "https://www.gstatic.com/firebasejs/11.1.0/firebase-firestore.js";
 import { loginData } from "./loginData.js";
 
-console.log(loginData)
 
 // catch basic elements
 const newArrivalsCont = document.getElementById("new-arrivals")
@@ -52,87 +51,6 @@ const displaynewArrivalProducts = async () => {
     alert("error in display data")
   }
 }
-
-window.addToFavs = async (buttonEl) => {
-  const productId = buttonEl.parentElement.id
-  const products = []
-  if (loginData.length < 1) {
-    alert("Please Sign-in to build a wishlist")
-  } else{
-    try {
-      const wishlistProds = await fetchData("newArrivalProducts", products)
-      const wishlist = loginData.wishlist
-      const userId = loginData.id
-      wishlistProds.forEach(async (item) => {
-        if(item.id == productId){
-          const findProd = wishlist.find((item) => item.id == productId)
-          if(!findProd){
-            // push item to wishlist
-            wishlist.push(item)
-            loginData.wishlist = wishlist
-            // and update the user's data in the localStorage
-            localStorage.setItem("loginData", JSON.stringify(loginData));
-
-            // Now update the firestore users db
-            try {
-              const userDocRef = doc(db, "users", userId);
-              await setDoc(userDocRef, { wishlist }, { merge:true })
-              alert("Item succesfully added to wishlist!")
-            } catch (error) {
-              alert("Problem updating the database with the product you added in wishlist")
-              console.log(error)
-            }
-          } else{
-            alert("Product is alraedy there in your wishlist")
-          }
-        }
-      })
-    } catch (error) {
-      alert("Sorry, could not add item to your wishlist, please try again")
-    }
-  }
-}
-
-window.addToCart = async (buttonEl) => {
-  const productId = buttonEl.parentElement.id
-  const products = []
-  if (loginData.length < 1) {
-    alert("Please Sign-in to build a cart")
-  } else{
-    try {
-      const cartProds = await fetchData("newArrivalProducts", products)
-      const cart = loginData.cart
-      const userId = loginData.id
-      cartProds.forEach(async (item) => {
-        if(item.id == productId){
-          const findProd = cart.find((item) => item.id == productId)
-          if(!findProd){
-            // push item to wishlist
-            cart.push(item)
-            loginData.cart = cart
-            // and update the user's data in the localStorage
-            localStorage.setItem("loginData", JSON.stringify(loginData));
-
-            // Now update the firestore users db
-            try {
-              const userDocRef = doc(db, "users", userId);
-              await setDoc(userDocRef, { cart }, { merge:true })
-              alert("Item succesfully added to cart!")
-            } catch (error) {
-              alert("Problem updating the database with the product you added in cart")
-              console.log(error)
-            }
-          } else{
-            alert("Product is alraedy there in your cart")
-          }
-        }
-      })
-    } catch (error) {
-      alert("Sorry, could not add item to your cart, please try again")
-    }
-  }
-}
-
 
 const displaynewInsiderProducts = async () => {
   try {
