+++ wishlist.html
@@ -66,5 +66,6 @@
 <script src="./scripts/navbar.js" type="module"></script>
 <script src="./scripts/footer.js"></script>
 <script src="./scripts/wishlist.js" type="module"></script>
+<script src="./scripts/cartWishlist.js" type="module"></script>
 
 </html>
